@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from enum import Enum
-from typing import Callable, Dict, List, Optional, Union
+from typing import Callable, Optional, Union
 
 from openfeature.exception import ErrorCode
 
@@ -18,19 +18,19 @@ class ProviderEvent(Enum):
 
 @dataclass
 class ProviderEventDetails:
-    flags_changed: Optional[List[str]] = None
+    flags_changed: Optional[list[str]] = None
     message: Optional[str] = None
     error_code: Optional[ErrorCode] = None
-    metadata: Dict[str, Union[bool, str, int, float]] = field(default_factory=dict)
+    metadata: dict[str, Union[bool, str, int, float]] = field(default_factory=dict)
 
 
 @dataclass
 class EventDetails(ProviderEventDetails):
     provider_name: str = ""
-    flags_changed: Optional[List[str]] = None
+    flags_changed: Optional[list[str]] = None
     message: Optional[str] = None
     error_code: Optional[ErrorCode] = None
-    metadata: Dict[str, Union[bool, str, int, float]] = field(default_factory=dict)
+    metadata: dict[str, Union[bool, str, int, float]] = field(default_factory=dict)
 
     @classmethod
     def from_provider_event_details(
